% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evacpath-package.R
\docType{package}
\name{evacpath}
\alias{evacpath}
\alias{evacpath-package}
\title{evacpath: Least-cost pedestrian evacuation modeling in R}
\description{
\code{evacpath} provides tools for modeling pedestrian evacuation distance and
travel time from hazard zones to safety areas using road-constrained least-cost
path analysis.
}
\details{
The package is organized as a set of modular functions. Typical workflows use
\code{prepare_tsunami_zones()} to create tsunami-specific hazard and escape zones,
\code{crop_roads_to_inner_extent()} and \code{make_road_aware_escape_zone()} to prevent
false escape points along artificial study-area boundaries while retaining
bridge and walkway corridors, \code{find_escape_points()} to identify candidate
exits, \code{make_conductance_surface()} to build a slope-based conductance surface,
and \code{run_evacpath()} to run the full workflow.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/el-cordero/evacpath}
  \item Report bugs at \url{https://github.com/el-cordero/evacpath/issues}
}

}
\author{
\strong{Maintainer}: Elvin Cordero \email{elvin.cordero1@upr.edu}

}
\keyword{package}
