% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{has_errors}
\alias{has_errors}
\title{Test whether diagnostics contain errors}
\usage{
has_errors(x, ...)
}
\arguments{
\item{x}{An object returned by \code{\link[=diagnose_evac_model]{diagnose_evac_model()}}.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Logical value indicating whether diagnostics contain at least one
error.
}
\description{
Test whether diagnostics contain errors
}
\examples{
dem <- terra::rast(nrows = 2, ncols = 2, vals = 1, crs = "EPSG:3857")
hazard <- terra::as.polygons(dem, dissolve = TRUE)
roads <- terra::vect(matrix(c(0, 1, 2, 1), ncol = 2, byrow = TRUE),
  type = "lines", crs = "EPSG:3857")
has_errors(diagnose_evac_model(hazard, roads, dem))
}
