% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_evacpath.R
\name{make_output_clip_area}
\alias{make_output_clip_area}
\title{Make an output clip area for evacuation polygons}
\usage{
make_output_clip_area(
  hazard_zone,
  roads_buffer,
  final_road_buffer_m = 3,
  clip_mode = c("hazard", "road_hazard", "hazard_plus_roads", "none")
)
}
\arguments{
\item{hazard_zone}{Hazard zone polygon.}

\item{roads_buffer}{Buffered roads.}

\item{final_road_buffer_m}{Additional road buffer for output clipping.}

\item{clip_mode}{One of \code{"hazard"}, \code{"road_hazard"}, \code{"hazard_plus_roads"}, or \code{"none"}.}
}
\value{
A \code{SpatVector} clip area or \code{NULL}.
}
\description{
Make an output clip area for evacuation polygons
}
