% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_time_polygons.R
\name{min_dist}
\alias{min_dist}
\title{Backward-compatible alias for minimum-distance calculation}
\usage{
min_dist(cs, origins, destinations)
}
\arguments{
\item{cs}{Conductance surface.}

\item{origins}{Origin points.}

\item{destinations}{Escape/safety points.}
}
\value{
A point \code{SpatVector} with distance values.
}
\description{
Backward-compatible alias for minimum-distance calculation
}
