% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_time_polygons.R
\name{calculate_min_dist}
\alias{calculate_min_dist}
\title{Calculate the minimum path distance from a list of least-cost paths}
\usage{
calculate_min_dist(lc_paths_list)
}
\arguments{
\item{lc_paths_list}{A list of least-cost path line vectors.}
}
\value{
Minimum non-zero, finite path distance.
}
\description{
Calculate the minimum path distance from a list of least-cost paths
}
